fix: Address review comments 2
Contributes to: #56
