+++ ui/src/Elements/Message/Message.scss
@@ -1,4 +1,7 @@
 .Message {
+    display: flex;
+    flex-direction: column;
+
     &__consumer-details {
         display: flex;
         flex-direction: row;
@@ -6,21 +9,25 @@
     }
 
     &__tile--producer {
-      padding: $spacing-05;
+        padding: $spacing-05;
+    }
+
+    &__tile--producer, &__tile--error {
+        height: 100%;
     }
 
     &__icon {
 
-      &--checkmark {
-        fill: $icon-02;
-      }
+        &--checkmark {
+            fill: $icon-02;
+        }
 
-      &--error {
-        fill: $support-01;
-      }
+        &--error {
+            fill: $support-01;
+        }
     }
 
     &--producer-first &__icon--checkmark {
-      fill: $support-02;
+        fill: $support-02;
     }
 }
